+++ examples/ios/BLEIntSDK/ContentView.swift
@@ -6,7 +6,7 @@
 //  Copyright © 2022 CocoaPods. All rights reserved.
 //
 
-import _2hire_BLEIntSDK
+import BLEIntSDK
 import SwiftUI
 import os.log
 
@@ -36,13 +36,12 @@ struct ContentView: View {
                     do {
                         let key = "Az8OqWnCYYyKCmuAJYrKUXDryu1HjpowPoRB8JkD5RRe"
 
-                        let commands: _2hire_BLEIntSDK.Commands = [
+                        let commands: BLEIntSDK.Commands = [
                             .Start:
                                 "ADqK3S1Ls8tu2NRha3/BKhdQOu3qCZgRXY1q/tDIKjzQhhcC3XqPlAXvN+Ct58ih8R2jKqkEZyldkKM0yvphMWgL4aLjE7rLnL+sKpTTofT+iriD+Ab3uhqrdVkSlHsfLGMfqAYvPvOHVhz8LoLoHK58EPsh/4PejhKBTpXPhbFpbO5dBllwdQJco9dSg4E38EDexbOmn5rcndwNWGLrGAY=",
                             .Stop: "AeUnVRMt7CF6dF6Ee3j3xDr+/MFVnZKN1L94oPVSiWHiSdV5/0TodK4zV3CkE9h67jQ8P4yym9vXqD6WoFt9flyxGWswV95d0RYX96w8W6UnCu3FLdWwgQYwAVCwmTU1Gqj3rXjzJs1fNgic0thn1cFqLtgNQefGm42zzivS48PBD2Mdk3LyFzzqJVw+HMCPG4x+sbkor8UcKhuI+0043Ig=",
                             .Noop:
                                 "AlbBTYOrZb2IO1kJd8uEnRgDsQE38yj6mRAiuQmrgW4wUOu0FPHKxZ3oN95tzf3zpKE1x2YcSqOavIdd3FvDwtm+GF9ISZ+/wHdwbfO4N0ODFJ7zgU9XVj+tflfyMYiLkbRQoeZMs4TR5eyIQ51o8FQoXKtdsesn8COZd6dtNPkcUL2+T6UdKyUHb9hLIs4U8ne7C3GmScsZiAzWlRKS3x4=",
-                            .Locate: "",
                             .EndSession:
                                 "/wz88sby9ft9Na1dp/Iq9Wwp+EQt04Xjh7rceWCVxs4Px8kXPyvjhsZuFHZ/usOhlV80Yh3X67VC2XyQl0pSbp29XCRmyAsFSblX8m60ABZXMFaQAysAHriDUh2sG4kesMAdrPxRDhRmeFA9yO5LHb4GYNalw/NC94EQm4BZAn//mgm/UEBTkfuJ2qj7qBY54La23Romht5O/Qa1bu3X5ds=",
                         ]
@@ -154,7 +153,6 @@ struct ContentView: View {
                 Text("Start").tag(CommandType.Start)
                 Text("Stop").tag(CommandType.Stop)
                 Text("Noop").tag(CommandType.Noop)
-                Text("Locate").tag(CommandType.Locate)
             }.pickerStyle(SegmentedPickerStyle())
                 .frame(width: 250)