+++ packages/cordova-plugin-bleintsdk/src/ios/BleIntSdkCordova.swift
@@ -1,9 +1,9 @@
 import Foundation
-import _2hire_BLEIntSDK
+import BLEIntSDK
 
 class BLEIntSDKCordova: CDVPlugin {
 
-    private static let sdkClient: _2hire_BLEIntSDK.Client = Client()
+    private static let sdkClient: BLEIntSDK.Client = Client()
 
     @objc(sessionSetup:)
     func sessionSetup(command: CDVInvokedUrlCommand) {
@@ -19,7 +19,7 @@ class BLEIntSDKCordova: CDVPlugin {
                 with: SessionData(
                     accessToken: accessToken,
                     publicKey: publicKey,
-                    commands: _2hire_BLEIntSDK.Commands.fromDictionary(dictionary: commands)
+                    commands: BLEIntSDK.Commands.fromDictionary(dictionary: commands)
                 )
             )
 
@@ -78,7 +78,7 @@ class BLEIntSDKCordova: CDVPlugin {
         Task.init {
             do {
                 guard let _command = command.argument(at: 0) as? String,
-                    let commandType = _2hire_BLEIntSDK.CommandType(rawValue: _command)
+                    let commandType = BLEIntSDK.CommandType(rawValue: _command)
                 else {
                     throw BridgeError.InvalidArgument
                 }
@@ -134,7 +134,7 @@ class BLEIntSDKCordova: CDVPlugin {
     }
 }
 
-extension _2hire_BLEIntSDK.CommandResponse: Encodable {
+extension BLEIntSDK.CommandResponse: Encodable {
     private enum CodingKeys: String, CodingKey {
         case success = "success"
         case additionalPayload = "payload"
@@ -154,11 +154,11 @@ extension _2hire_BLEIntSDK.CommandResponse: Encodable {
     }
 }
 
-extension _2hire_BLEIntSDK.Commands {
+extension BLEIntSDK.Commands {
     fileprivate static func fromDictionary(dictionary: NSDictionary) -> Self {
         var result: Self = [:]
 
-        _2hire_BLEIntSDK.CommandType.allCases.forEach {
+        BLEIntSDK.CommandType.allCases.forEach {
             if let value = dictionary[$0.rawValue] as? String {
                 result[$0] = value
             }