+++ packages/sdk/android/core/src/main/java/io/twohire/bleintsdk/client/CommandType.kt
@@ -4,7 +4,6 @@ enum class CommandType(val rawValue: String) {
     Start("start"),
     Stop("stop"),
     Noop("noop"),
-    Locate("locate"),
     EndSession("end_session");
 
     companion object {