+++ packages/sdk/ios/K1/Classes/Bridge/BridgeToC.swift
@@ -7,7 +7,7 @@
 
 
 import CryptoKit 
-import _2hire_secp256k1
+import secp256k1
 
 internal final class Bridge {