+++ packages/sdk/ios/K1/Classes/K1/ECDSA/ECDSASignature+From+DER.swift
@@ -7,7 +7,7 @@
 
 
 // Bridge to C
-import _2hire_secp256k1
+import secp256k1
 
 extension Bridge {
     /// Initializes ECDSASignature from the DER representation.