+++ packages/sdk/ios/K1/Classes/K1/Keys/PublicKey/PublicKey/PublicKey+Bridge+To+C.swift
@@ -7,7 +7,7 @@
 
 
 // Bridge to C
-import _2hire_secp256k1
+import secp256k1
 import CryptoKit
 import Foundation