+++ packages/sdk/ios/K1/Classes/K1/Keys/PublicKey/Wrapped/PublicKey+Wrapped+Bridge+To+C.swift
@@ -6,7 +6,7 @@
 //
 
 
-import _2hire_secp256k1
+import secp256k1
 
 extension Bridge {