+++ packages/sdk/ios/core/BLEIntSDK/Classes/Protocol/ProtocolManager.swift
@@ -5,7 +5,7 @@
 //
 
 import Foundation
-import _2hire_K1
+import K1
 import os.log
 
 public struct ProtocolResponse {